productive/flaskClient/requirements.txt
Flask
flask-socketio
psycopg2-binary
